+++ authorization/organization/organization_model_service_blackbox_test.go
@@ -1,11 +1,10 @@
-package model_test
+package organization
 
 import (
 	"testing"
 
 	"github.com/fabric8-services/fabric8-auth/account"
-	"github.com/fabric8-services/fabric8-auth/authorization/common"
-	"github.com/fabric8-services/fabric8-auth/authorization/model"
+	"github.com/fabric8-services/fabric8-auth/authorization/organization/common"
 	"github.com/fabric8-services/fabric8-auth/authorization/resource"
 	"github.com/fabric8-services/fabric8-auth/authorization/role"
 	"github.com/fabric8-services/fabric8-auth/gormtestsupport"
@@ -24,7 +23,7 @@ type organizationModelServiceBlackBoxTest struct {
 	resourceRepo     resource.ResourceRepository
 	resourceTypeRepo resource.ResourceTypeRepository
 	roleRepo         role.RoleRepository
-	orgModelService  model.OrganizationModelService
+	orgModelService  OrganizationModelService
 }
 
 func TestRunOrganizationModelServiceBlackBoxTest(t *testing.T) {
@@ -40,7 +39,7 @@ func (s *organizationModelServiceBlackBoxTest) SetupTest() {
 	s.resourceTypeRepo = resource.NewResourceTypeRepository(s.DB)
 	s.roleRepo = role.NewRoleRepository(s.DB)
 
-	s.orgModelService = model.NewOrganizationModelService(s.DB, &test.TestRepositories{
+	s.orgModelService = NewOrganizationModelService(s.DB, &test.TestRepositories{
 		FIdentityRepository:     s.identityRepo,
 		FIdentityRoleRepository: s.identityRoleRepo,
 		FResourceRepository:     s.resourceRepo,
