+++ authorization/role_assignment_service.go
@@ -3,23 +3,22 @@ package authorization
 import (
 	"context"
 	"github.com/fabric8-services/fabric8-auth/application"
-	"github.com/fabric8-services/fabric8-auth/authorization/model"
+	"github.com/fabric8-services/fabric8-auth/authorization/assignment"
 	"github.com/fabric8-services/fabric8-auth/authorization/role"
 	"github.com/fabric8-services/fabric8-auth/errors"
 	"github.com/fabric8-services/fabric8-auth/log"
 )
 
 type RoleAssignmentService interface {
-	//Assign()
 	ListByResource(ctx context.Context, resourceID string) ([]role.IdentityRole, error)
 }
 
 type RoleAssignmentServiceImpl struct {
-	modelService model.RoleAssignmentModelService
+	modelService assignment.RoleAssignmentModelService
 	db           application.DB
 }
 
-func NewRoleAssignmentService(modelService model.RoleAssignmentModelService, db application.DB) *RoleAssignmentServiceImpl {
+func NewRoleAssignmentService(modelService assignment.RoleAssignmentModelService, db application.DB) *RoleAssignmentServiceImpl {
 	return &RoleAssignmentServiceImpl{modelService: modelService, db: db}
 }
