+++ main.go
@@ -30,7 +30,8 @@ import (
 	"github.com/fabric8-services/fabric8-auth/token/link"
 
 	"github.com/fabric8-services/fabric8-auth/authorization"
-	"github.com/fabric8-services/fabric8-auth/authorization/model"
+	"github.com/fabric8-services/fabric8-auth/authorization/assignment"
+	"github.com/fabric8-services/fabric8-auth/authorization/organization"
 	"github.com/goadesign/goa"
 	"github.com/goadesign/goa/logging/logrus"
 	"github.com/goadesign/goa/middleware"
@@ -183,7 +184,7 @@ func main() {
 	loginCtrl := controller.NewLoginController(service, loginService, tokenManager, config)
 	app.MountLoginController(service, loginCtrl)
 
-	roleAssignmentModelService := model.NewRoleAssignmentModelService(db, appDB)
+	roleAssignmentModelService := assignment.NewRoleAssignmentModelService(db, appDB)
 	roleAssignmentService := authorization.NewRoleAssignmentService(roleAssignmentModelService, appDB)
 	resourceRoleCtrl := controller.NewResourceRolesController(service, appDB, roleAssignmentService)
 	app.MountResourceRolesController(service, resourceRoleCtrl)
@@ -254,7 +255,7 @@ func main() {
 	app.MountResourceController(service, resourcesCtrl)
 
 	// Mount "organizations" controller
-	organizationModelService := model.NewOrganizationModelService(db, appDB)
+	organizationModelService := organization.NewOrganizationModelService(db, appDB)
 	organizationService := authorization.NewOrganizationService(organizationModelService, appDB)
 	organizationCtrl := controller.NewOrganizationController(service, appDB, organizationService)
 	app.MountOrganizationController(service, organizationCtrl)
